refactor: created separate transform.js for record and conversion use case
