+++ src/cdk/v2/destinations/the_trade_desk/rtWorkflow.yaml
@@ -1,6 +1,6 @@
 bindings:
   - name: processRouterDest
-    path: ./utils
+    path: ./transform
 
 steps:
   - name: validateInput
